Moved history tests to their own file. Simplified adding user history entries a bit
